review: lift TLS adapter to shared util and apply it in api_source_manager
Address PR review item 1 (medium):

The first revision claimed `api_source_manager.py` was fixed for both Xray
findings (no `verify=False` literal AND TLS 1.2 minimum), but only the
literal was addressed — the call site was still `requests.request(...)`
with no declared minimum TLS version. AST-based scanners would still fire.

Changes:
- Lift `TLSMinV1_2Adapter` and a new `make_secure_session(ssl_verify=...)`
  factory into `core/utils/http_utils.py`. Single canonical place for the
  SSLContext + adapter pattern; future API processors that face this same
  Xray finding can adopt it without copy-paste.
- `argocd_api_processor.py`: drop the local adapter class and the manual
  Session wiring; `__init__` now calls `make_secure_session(...)` directly.
  Behavior identical (smoke-tested: ctx.minimum_version=TLSv1_2 in both
  modes; CERT_NONE/check_hostname=False when ssl_verify=False;
  CERT_REQUIRED/check_hostname=True when ssl_verify=True).
- `api_source_manager.execute_http_request`: replaces the module-level
  `requests.request(...)` with a `make_secure_session(ssl_verify=...)`
  session per call. Session is local to the function (no reuse benefit
  for one-shot HTTP requests; cost is negligible relative to network
  round-trip). Now the file declares both the minimum TLS version and
  the verification posture explicitly.

Also addresses review item 3 (doc):

The "backward compatibility" comment on the `bool(...)` rewrite misframed
the intent — the form-field `default_value=BoolValue(value=True)` means
new UI-created connectors start at True; the False fallback applies only
to connectors persisted before the `ssl_verify` field existed in the
schema. Re-worded the comment per the reviewer's suggestion.

Items 4 and 5 are nits; left as-is per reviewer's "your call" / common
case (`ssl_verify` keyword name is distinct from `verify=`, so the
default literal `False` does not match Xray's `verify=False` rule).
