+++ drdroid_debug_toolkit/core/integrations/source_api_processors/argocd_api_processor.py
@@ -1,46 +1,10 @@
 import logging
-import ssl
-
-import requests
-from requests.adapters import HTTPAdapter
-from urllib3.poolmanager import PoolManager
 
 from core.integrations.processor import Processor
 from core.settings import EXTERNAL_CALL_TIMEOUT
+from core.utils.http_utils import make_secure_session
 
 logger = logging.getLogger(__name__)
-
-
-class _TLSMinV1_2Adapter(HTTPAdapter):
-    """HTTPAdapter that pins outbound HTTPS to TLS 1.2 or higher.
-
-    Matches the system OpenSSL default on the runtime image; declared explicitly
-    so static analysers see the minimum-version requirement. The adapter also
-    carries the verification posture so the SSLContext built here is internally
-    consistent — relying on `session.verify=False` to override a CERT_REQUIRED
-    context after the fact has historically been fragile across urllib3
-    versions.
-    """
-
-    def __init__(self, ssl_verify=True, **kwargs):
-        self._ssl_verify = ssl_verify
-        super().__init__(**kwargs)
-
-    def init_poolmanager(self, connections, maxsize, block=False, **pool_kwargs):
-        ctx = ssl.create_default_context()
-        ctx.minimum_version = ssl.TLSVersion.TLSv1_2
-        if not self._ssl_verify:
-            # Order matters: check_hostname must be cleared before verify_mode
-            # can be set to CERT_NONE.
-            ctx.check_hostname = False
-            ctx.verify_mode = ssl.CERT_NONE
-        pool_kwargs["ssl_context"] = ctx
-        self.poolmanager = PoolManager(
-            num_pools=connections,
-            maxsize=maxsize,
-            block=block,
-            **pool_kwargs,
-        )
 
 
 class ArgoCDAPIProcessor(Processor):
@@ -51,9 +15,7 @@ class ArgoCDAPIProcessor(Processor):
         # existing customers running self-signed-cert ArgoCD instances continue
         # to work after upgrade. Pass ssl_verify=True to enable validation.
         self._ssl_verify = ssl_verify
-        self._session = requests.Session()
-        self._session.mount("https://", _TLSMinV1_2Adapter(ssl_verify=self._ssl_verify))
-        self._session.verify = self._ssl_verify
+        self._session = make_secure_session(ssl_verify=self._ssl_verify)
 
     def _auth_headers(self):
         return {'Authorization': f'Bearer {self.__token}'}
